requirements.txt
flask
pymongo
pymongo
pydantic
pydantic[email]
qrcode
python-dotenv
gunicorn
gevent
flask_socketio